Add getResponsePageByPageId function
There’s fewer complications here than in the title function, because
normalization and redirects don’t apply. And in the formatversion=1
version, we can skip the loop and rely on the fact that the page ID is
used as the key.

We compare the page IDs as strings, even if they’re returned as numbers
by the API, because numbers can theoretically overflow in JS (though any
real MediaWiki install is far, far away from that point), and I want the
library to be prepared for the API potentially returning strings in the
future.