Add queryPartialPageByTitle function
This is the first out of a planned set of three functions: “partial”
makes a single request and returns the page as it is in the response;
“incremental” follows continuation until a batchcomplete response and
yields the partial pages; finally, “full” merges all these partial pages
back into one full page. Most people will want to use “full”;
“incremental” could be useful if you don’t trust the merging logic (not
having tried to write it yet, I’m not sure I would trust it myself ^^),
and “partial” isn’t really something you want but it’s something to
start with. And then after that we’ll add the same three functions for
page IDs and revision IDs instead of titles, too, and finally a function
that instead follows a generator and yields full pages from it.