Prevent mixing titles and generator
When titles= (or pageids= or revids=) are used with generator=, they
lose their ordinary meaning and instead become the input to the
generator (e.g. generator=links will generate the outgoing links from
those pages). The CombiningSession class already takes care that the
parameters aren’t inappropriately mixed accidentally, but we also need
to watch out here that it doesn’t happen when we add titles to the
existing params supplied by the user – throw an exception in this case.