Add queryFullPageByTitle function
This is one of the main functions people will want to use, and the first
one to merit inclusion in the README – everything up to here was more
preparation for this function than useful in its own right. I’ve tested
it with the longest current enwiki page (2022 United States House of
Representatives elections) and all the props that worked without error
(too many to list), and it didn’t crash and the resulting page looked
sensible, so I think the merging logic might work correctly.