Fall back from pageid to title for missing pages
It’s possible for a generator to generate missing pages, e.g. the search
generator can generate files that don’t exist on the local wiki when
InstantCommons is enabled. In this case, several pages in the result may
have the pageid 0, but they can still be distinguished by their title.