Allow some different values when merging responses
Testing with generator=search revealed that search results aren’t
entirely stable during continuation, which made the previous
implementation crash: the same page could have slightly different index
values between responses (off by between zero and five or so), and we
required the value to be the same. Relax this requirement a bit, but
also make it configurable: by default, allow strings and numbers (but
not booleans) to vary between responses, and use the earlier value; a
custom callback can make a different decision, and potentially make
other modifications to the object too (e.g. attach both values under a
custom symbol). I think this is an acceptable compromise.