Implement queryFullPageByTitle() using requestAndContinueReducingBatch()
Fixes #2 (requestAndContinueReducingBatch() hides truncatedresult
warnings). queryPartialPageByTitle() should never hide those warnings;
for queryIncrementalPageByTitle() I’d say it’s arguable, but that
function can’t be implemented using requestAndContinueReducingBatch(),
so for now it still shows the warnings. (I considered removing the
function, but the queryFullPageByTitle() documentation actually points
out a valid use case for it. Maybe there’ll be a separate function to
enumerate revisions of a page later, but for now the function stays.)