Add getResponseRevisionByRevisionId function
Fixes #1. The MediaWiki API returns bad revision IDs under a separate
"badrevids" key, with the individual revisions having a "missing" key
(MediaWiki patch contributed by yours truly; backported to 1.37, 1.36
and 1.35 by Reedy; we add the key if it’s not there, for responses by
older MediaWiki versions). We return this object, which is consistent
with the page functions (return null only on param mismatch).

I considered adding some way for users to specify the title or page ID
of the page that the revision belongs to, to avoid looping over all
revisions of all pages in the response, but concluded that it’s not
necessary: prop=revisions refuses to add more than one revision to more
than one page (rvlimit>1 can be used with generator, but only if the
generator is limited to one page per response), so there’s no point in
finding the right page by title: either there’s a single page in the
response, or each page only contains only one revision, so it doesn’t
make a difference whether you’re comparing each page’s title or each
revision’s ID. Finding the right page by page ID could help, but only
for formatversion=1 responses (where you can use the page ID as the
object key), and that’s not a case I want to optimize the whole function
interface for (though we make use of it in getResponsePageByPageId()).