Add query-by-page-ID functions
The partial/incremental/full part is just like for the by-title
functions. However, the page ID can also be a number instead of a
string, which means that makeParamsWithPageId() is a bit more involved
than makeParamsWithTitle(); because (out of sheer stubbornness) I want
this code to work correctly for page IDs that don’t fit in a JS number,
we turn all page IDs into strings in that function.

Another set of by-revision-ID functions will also follow, but there the
function split will be different from partial/incremental/full, since
the API never returns partial revisions.