Add query-by-revision-ID functions
Only two functions rather than three as for pages: the API never returns
incomplete revisions (a revision is either returned fully or not at
all), so the “partial” and “incremental” functions are merged into a
single “potential” function that yields nulls until the revision is
found, then stops.

Currently, queryPotentialRevisionByRevisionId is based on
requestAndContinue, while queryFullRevisionByRevisionId is based on
requestAndContinueReducingBatch. Neither is ideal – the former means
warnings about truncated results are still shown even though they’re not
necessary in our case, while the latter means that we potentially keep
following continuation even after we found the revision we’re interested
in. This will be resolved with the new dropTruncatedResultWarning option
introduced in the upcoming m3api release (probably 0.6.1 – see commit
lucaswerkmeister/m3api@2138581290), but I thought it would be useful to
commit this separately first.