Attach revision’s page using custom symbol
Introduce a symbol which we use to decorate a revision object, returned
by the various revision functions, with its surrounding page. This is
expected to be useful in some situations where you want information
about the page as well – especially in the upcoming queryFullRevisions
(see #3), but also in the existing functions.

The symbol is not enumerable, so that it doesn’t pollute e.g.
console.log() output and generally doesn’t disturb anyone who’s not
interested in it; but it is writable and configurable, so that users who
want to change something about it can do so.

In the tests, a bunch of .equal() assertions need to change to .eql(),
since the returned object is no longer the same instance as the input
revision. For practical usage outside of tests, this should be fine.
(You could even argue that all the assertions, including for the “page”
functions, should use .eql() to start with, I think.)