Take converted titles into account
And add a first integration test that actually makes requests against a
real MediaWiki installation. Use Serbian Wikipedia for testing (if
T299975 was done, we could use Pig Latin on Beta English Wikipedia
instead, but it’s not done yet), with a title that exhibits all layers
of title modification and will hopefully be relatively stable: Latin
“belgrad” is normalized to “Belgrad”, converted to Cyrillic “Белград”,
then redirected to “Београд”.

Fixes #7.