Update integration test page
In 1e7fda2801, I had hoped that the test case would be relatively
stable, but it was broken in January 2025 by making srwiki „Белград” a
disambiguation instead of a redirect [1]. Pick a different example. The
goal is the same: to exhibit all layers of title modification, in that
nemačka is normalized to Nemačka, converted to Cyrillic Немачка, then
redirected to Њемачка (note the added soft sign on the first letter).

[1]: https://sr.wikipedia.org/w/index.php?title=%D0%91%D0%B5%D0%BB%D0%B3%D1%80%D0%B0%D0%B4&diff=prev&oldid=28964978