Add optional protection against infinite continuation
Similar to the previous commit, this aims to protect against cases where
queryFullPages() and queryFullRevisions() may follow continuation
forever without yielding any results back to the caller; this time,
we’re worried about API parameters that will rarely or never return
results, but which can’t be detected by static analysis as in the
previous commit. (An example would be querying www.wikidata.org with
generator=allrevisions and garvnamespace=122 – the Query namespace,
which has no pages in it.) The new function maxEmptyResponses() can be
used to throw an exception after too many consecutive empty responses.

To implement this, two new request options are added, and
maxEmptyResponses() returns an object with those options which the
caller should spread into the options (either when creating the session
or when calling queryFullPages() / queryFullRevisions()). The options
themselves are part of the internal interface (we follow m3api’s semver
now) – this way, I can change the way the options work later without
breaking anything for normal users. (The options should not be used for
anything else; in theory they can reorder, add or drop pages/revisions
from the results to be yielded, but this would result in very confusing
code, as it’s a form of highly non-local influence if set up at session
creation time.)