Let maxEmptyResponses() distinguish separate calls
Previously, one maxEmptyResponses() function just had one counter of
consecutive empty responses; if you used it for multiple requests (e.g.
if you had specified it once when creating the session), that counter
would get mixed up between different calls to queryFullPages() or
queryFullRevisions() and potentially not do the intended thing. Fix that
by calling the callback with a state object which those functions keep
track of throughout the continuation. For good measure, also pass the
session, params and options into the callback, though we don’t use them
here.

The test for this is a carefully choreographed ballet of promises that
we resolve and await in *exactly* the right order; to convince myself
that it worked and tested the right thing, I followed proper TDD for
once and implemented the test before the feature (the test was the
*much* more difficult thing anyway). There were some tricky spots where
I needed to figure out why it wasn’t behaving as expected (e.g. it’s
very important that we “drive” the generator by calling .next() quickly
and then await the returned promise later, rather than directly awaiting
.next() later!), but now it works and I hope it’s not *too* horribly
unreadable. The one thing we annoyingly can’t assert is that the promise
(the one returned by the functions under test, not our own deferred())
hasn’t settled before we expect it to, but we’ll just have to live with
that, I think – I can’t think of a way to do that.