Add support for undoing maxEmptyResponses()
This way, you can e.g. specify maxEmptyResponses() in your default
options but then override it for a more specific request (though I
struggle to imagine when you might ever want *no* limit at all). I think
prior to this, maxEmptyResponses( Infinity ) also would’ve worked, but
it would’ve been less efficient.

This is also important for the future – I want to turn on
maxEmptyResponses() by default in the next major release, and then we
need a way for users to undo it.