Add planned breaking change to CHANGES.md
systemd does this in their release notes, and I think for m3api
(extension packages) it’s a reasonable, low-effort place to put this
stuff so I don’t forget it. (A Phabricator task feels like too much
overhead.) I don’t think it needs to be shown to users though, so just
comment it out for now.